requirements.txt
streamlit
ultralytics
python-dotenv
mlflow
docker
pillow
PyYAML
roboflow
opencv-python-headless
torch
torchvision